+++ protosocket-rpc/src/client/mod.rs
@@ -2,7 +2,10 @@ mod configuration;
 mod reactor;
 mod rpc_client;
 
-pub use configuration::{connect, Configuration, StreamConnector, TcpStreamConnector};
+pub use configuration::{
+    connect, Configuration, StreamConnector, TcpStreamConnector, UnverifiedTlsStreamConnector,
+    WebpkiTlsStreamConnector,
+};
 pub use rpc_client::RpcClient;
 
 pub use reactor::completion_streaming::StreamingCompletion;
